# api/openapi/subscription-api.yaml
openapi: 3.0.3
info:
  title: Subber Subscription API
  version: 1.0.0
paths:
  /api/subscribe:
    post:
      summary: Subscribe an email to GitHub repository releases.
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubscribeRequest'
      responses:
        "200":
          description: Subscription created and confirmation email requested.
        "400":
          description: Invalid request.
        "404":
          description: Repository does not exist on GitHub.
        "409":
          description: Email is already subscribed to this repository.
  /api/subscriptions/:
    get:
      summary: List confirmed subscriptions for an email.
      security:
        - ApiKeyAuth: []
      parameters:
        - name: email
          in: query
          required: true
          schema:
            type: string
            format: email
      responses:
        "200":
          description: Confirmed subscriptions.
  /api/confirm/{token}:
    get:
      summary: Confirm a subscription.
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Subscription confirmed.
        "404":
          description: Token not found.
  /api/unsubscribe/{token}:
    get:
      summary: Unsubscribe by token.
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Subscription removed.
        "404":
          description: Token not found.
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
    SubscribeRequest:
      type: object
      required:
        - email
        - repo
      properties:
        email:
          type: string
          format: email
        repo:
          type: string
          example: gin-gonic/gin